requirements.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses==0.6
dm-tree==0.1.8
docker-pycreds==0.4.0
flatbuffers==1.12
fsspec==2024.10.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googledrivedownloader @ file:///home/dang/code/continual_learning_framework/assets/googledrivedownloader-0.4-py2.py3-none-any.whl#sha256=278dc5fa6ad7027786d648977169a20ab3ba0b0985fd33965443e8585874ed87
grpcio==1.64.0
h5py==3.11.0
huggingface-hub==0.26.1
idna==3.7
importlib_metadata==7.1.0
joblib==1.4.2
keras==2.9.0
Keras-Preprocessing==1.1.2
kornia==0.7.2
kornia_rs==0.1.3
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.1
pandas==2.0.3
Pillow==9.3.0
platformdirs==4.2.2
protobuf==3.19.6
psutil==5.9.8
pyaml==24.4.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.10.1
sentry-sdk==2.2.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
termcolor==2.4.0
threadpoolctl==3.1.0
tiktoken==0.7.0
tokenizers==0.20.1
torch==1.12.0+cu102
torchaudio==0.12.1+cu102
torchvision==0.13.1+cu102
transformers==4.45.2
triton==2.3.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.1
umap-learn==0.1.5
urllib3==2.2.1
wandb==0.17.0
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.18.2
